requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
Django==4.0.1
fonttools==4.29.0
gunicorn>=20.1.0,<20.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
protobuf==3.19.3
PyMuPDF==1.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stanford-openie==1.3.0
stanfordnlp==0.2.0
torch==1.10.1
tqdm==4.62.3
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wget==3.2
wordcloud==1.8.1
